requirements.txt
Flask
gunicorn
celery[sqs]
boto3
python-dotenv